fix track websocket test
