+++ test/ioserver.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package test
 
 import (
